reup

+++ database/migrations/2024_02_15_055800_create_user_groups_table.php
@@ -15,7 +15,7 @@ class CreateUserGroupsTable extends Migration
     {
         Schema::create('user_groups', function (Blueprint $table) {
             $table->id();
-            $table->integer('u_id')->unsigned()->nullable();
+            $table->bigInteger('u_id')->unsigned()->nullable();
             $table->bigInteger('g_id')->unsigned()->nullable();
             $table->foreign('u_id')->references('u_id')->on('users')->onUpdate('cascade')->onDelete('cascade');
             $table->foreign('g_id')->references('g_id')->on('groups')->onDelete('cascade')->onDelete('cascade');
